restructuring sparsity
