+++ examples/offline_inference_semi_structured_sparse.py
@@ -5,7 +5,7 @@ os.environ["CUDA_VISIBLE_DEVICES"] = "0" # Use only cuda:0
 
 model = LLM(
     "nm-testing/zephyr-50sparse-24", 
-    sparsity="semi_structured_sparse_w16a16",   # If left off, model will be loaded as dense
+    sparsity="sparse_w16a16", # semi_structured_sparse_w16a16   # If left off, model will be loaded as dense
     enforce_eager=True,         # Does not work with cudagraphs yet
     dtype="float16",
     tensor_parallel_size=1,
